fix(#2745 R3): Dual monotonic floor + existing-task authority + de-monolith (t-…58)
Addresses root R2 exact-head REJECT (both P0 findings) at 363228ef.

Finding 1 (P0) — two-observation no-weaken bypass: reconcile_review_class mapped
(Dual,Unresolved)->Unresolved then (Unresolved,Single)->Single, so a stale/missing
obs laundered the Dual floor and a later Single re-arm downgraded a 2-reviewer gate
to 1. FIX: Dual is a MONOTONIC FLOOR — (Dual,_)->Dual for any ordinary watch; an
intentional downgrade is a separate audited reset. Production SEQUENCE RED
r3_dual_floor_survives_unresolved_then_single_sequence_2745 (through record_ci_result,
with one existing VERIFIED at head) + updated transition matrix.

Finding 2 (P0) — existing-task authority not enforced: raw_review_class fell back to
args.review_class regardless of task_id, so an untagged existing task passed by
supplying send.review_class (the task stayed authority-less). FIX: the send-arg
fallback is now AUTO-CREATE-only (empty task_id); an existing task's metadata is the
SOLE authority via resolve_existing_task_review_class — a send arg / second_reviewer
is consistency-evidence only (can neither fill a missing-metadata gap nor contradict;
invalid/missing task lookup rejects). Handler REDs: untagged task + send class ->
Unspecified + no watch/bind/send; tagged task + contradictory send -> Mismatch + no
watch. Unit truth-table for the resolver.

De-monolith (root/reviewer-flagged Coverage failure): pr_state/mod.rs test module
(~2581 LOC) re-homed to sibling `tests.rs` via `#[path]` (established pattern);
production mod.rs 3819->1328 LOC, de-grandfathered from src_file_size_invariant.

Bundled call-site updates for the fail-closed authority: e2e_workflow + teardown
create their dispatched task WITH review_class metadata (a send arg can no longer
fill it); dispatch_hook lease/watch tests seed a real tagged task.

Verified (real git + tray, CI-exact): pr_state 99; full main-bin 5171/0; e2e +
teardown pass; src_file_size_invariant + file_size_invariant pass; clippy
(--features tray -D warnings) exit 0; CI fmt exit 0.

Agend-Agent: archfix-opus-2
Agend-Task: t-20260712092041221831-40783-58
Agend-Branch: fix/durable-review-class-authority
Agend-Issued-At: 2026-07-12T09:33:36.676219+00:00
