+++ tests/teardown_completeness_regression.rs
@@ -216,7 +216,9 @@ fn run_scenario(home: &Path) {
         &h,
         "lead",
         "task",
-        json!({"action": "create", "title": "do work"}),
+        // #2745 R3: the durable review_class authority lives on the TASK (a send arg
+        // can no longer fill a missing one), so the dispatched task is created tagged.
+        json!({"action": "create", "title": "do work", "review_class": "single"}),
     );
     let tid = created["result"]["id"].as_str().unwrap_or("").to_string();
     let _ = mcp(
